support testing with different datasets
