+++ tests/e2e/vLLM/configs/actorder/w4a16_actorder_group.yaml
@@ -2,6 +2,7 @@ cadence: "nightly"
 test_type: "regression"
 model: Qwen/Qwen2-0.5B-Instruct
 recipe: tests/e2e/vLLM/recipes/actorder/recipe_w4a16_actorder_group.yaml
-dataset_id: HuggingFaceH4/ultrachat_200k
-dataset_split: train_sft
+dataset_id: openai/gsm8k
+dataset_config: main
+dataset_split: train
 scheme: W4A16
